+++ components/blog/ArticleLink.tsx
@@ -1,0 +1,9 @@
+import Link from 'next/link'
+
+export const ArticleLink = (props: { id: number; title: string }) => (
+  <li key={props.id}>
+    <Link as={`/blog/${props.id}`} href={`/article?id=${props.id}`}>
+      <a>{props.title}</a>
+    </Link>
+  </li>
+)
