--- models/Content.tsx
@@ -1,5 +1,0 @@
-export interface Content {
-  id: number
-  name: string
-  imageUrl: string
-}
